+++ src/main/scala/scalaz/nio/channels/AsynchronousChannel.scala
@@ -1,0 +1,223 @@
+package scalaz.nio.channels
+
+import java.lang.{ Integer => JInteger, Long => JLong, Void => JVoid }
+import java.net.{ SocketAddress, SocketOption }
+import java.nio.channels.{
+  AsynchronousByteChannel => JAsynchronousByteChannel,
+  AsynchronousChannelGroup => JAsynchronousChannelGroup,
+  AsynchronousServerSocketChannel => JAsynchronousServerSocketChannel,
+  AsynchronousSocketChannel => JAsynchronousSocketChannel,
+  CompletionHandler => JCompletionHandler
+}
+
+import scalaz.nio.ByteBuffer
+import scalaz.nio.channels.AsynchronousChannel._
+import scalaz.zio.{ Async, ExitResult, IO }
+import scalaz.{ IList, Maybe }
+
+import scala.concurrent.duration.Duration
+
+class AsynchronousByteChannel(private val channel: JAsynchronousByteChannel) {
+
+  /**
+   *  Reads data from this channel into buffer, returning the number of bytes
+   *  read, or -1 if no bytes were read.
+   */
+  final def read(b: ByteBuffer): IO[Exception, Int] =
+    wrap[Unit, JInteger](h => channel.read(b.buffer, (), h)).map(_.toInt)
+
+  /**
+   *  Reads data from this channel into buffer, returning the number of bytes
+   *  read, or -1 if no bytes were read.
+   */
+  final def read[A](b: ByteBuffer, attachment: A): IO[Exception, Int] =
+    wrap[A, JInteger](h => channel.read(b.buffer, attachment, h)).map(_.toInt)
+
+  /**
+   *  Writes data into this channel from buffer, returning the number of bytes written.
+   */
+  final def write(b: ByteBuffer): IO[Exception, Int] =
+    wrap[Unit, JInteger](h => channel.write(b.buffer, (), h)).map(_.toInt)
+
+  /**
+   *  Writes data into this channel from buffer, returning the number of bytes written.
+   */
+  final def write[A](b: ByteBuffer, attachment: A): IO[Exception, Int] =
+    wrap[A, JInteger](h => channel.write(b.buffer, attachment, h)).map(_.toInt)
+
+  /**
+   * Closes this channel.
+   */
+  final def close: IO[Exception, Unit] =
+    IO.syncException(channel.close())
+
+}
+
+class AsynchronousServerSocketChannel(private val channel: JAsynchronousServerSocketChannel) {
+
+  /**
+   * Binds the channel's socket to a local address and configures the socket
+   * to listen for connections.
+   */
+  final def bind(address: SocketAddress): IO[Exception, Unit] =
+    IO.syncException(channel.bind(address)).void
+
+  /**
+   * Binds the channel's socket to a local address and configures the socket
+   * to listen for connections, up to backlog pending connection.
+   */
+  // TODO wrap `SocketAddress`
+  final def bind(address: SocketAddress, backlog: Int): IO[Exception, Unit] =
+    IO.syncException(channel.bind(address, backlog)).void
+
+  // TODO wrap `SocketOption[T]?`
+  final def setOption[T](name: SocketOption[T], value: T): IO[Exception, Unit] =
+    IO.syncException(channel.setOption(name, value)).void
+
+  /**
+   * Accepts a connection.
+   */
+  final def accept: IO[Exception, AsynchronousSocketChannel] =
+    wrap[Unit, JAsynchronousSocketChannel](h => channel.accept((), h))
+      .map(AsynchronousSocketChannel(_))
+
+  /**
+   * Accepts a connection.
+   */
+  final def accept[A](attachment: A): IO[Exception, AsynchronousSocketChannel] =
+    wrap[A, JAsynchronousSocketChannel](h => channel.accept(attachment, h))
+      .map(AsynchronousSocketChannel(_))
+
+  /**
+   * The `SocketAddress` that the socket is bound to,
+   * or the `SocketAddress` representing the loopback address if
+   * denied by the security manager, or `Maybe.empty` if the
+   * channel's socket is not bound.
+   */
+  final def localAddress: IO[Exception, Maybe[SocketAddress]] =
+    IO.syncException(Maybe.fromNullable(channel.getLocalAddress))
+
+  /**
+   * Closes this channel.
+   */
+  final def close: IO[Exception, Unit] =
+    IO.syncException(channel.close())
+
+}
+
+object AsynchronousServerSocketChannel {
+
+  def apply(): IO[Exception, AsynchronousServerSocketChannel] =
+    IO.syncException(JAsynchronousServerSocketChannel.open())
+      .map(new AsynchronousServerSocketChannel(_))
+
+  def apply(
+    channelGroup: AsynchronousChannelGroup
+  ): IO[Exception, AsynchronousServerSocketChannel] =
+    IO.syncException(
+        JAsynchronousServerSocketChannel.open(channelGroup.jChannelGroup)
+      )
+      .map(new AsynchronousServerSocketChannel(_))
+}
+
+class AsynchronousSocketChannel(private val channel: JAsynchronousSocketChannel)
+    extends AsynchronousByteChannel(channel) {
+
+  final def bind(address: SocketAddress): IO[Exception, Unit] =
+    IO.syncException(channel.bind(address)).void
+
+  // TODO wrap `SocketOption[T]?`
+  final def setOption[T](name: SocketOption[T], value: T): IO[Exception, Unit] =
+    IO.syncException(channel.setOption(name, value)).void
+
+  final def shutdownInput: IO[Exception, Unit] =
+    IO.syncException(channel.shutdownInput()).void
+
+  final def shutdownOutput: IO[Exception, Unit] =
+    IO.syncException(channel.shutdownOutput()).void
+
+  final def remoteAddress: IO[Exception, SocketAddress] =
+    IO.syncException(channel.getRemoteAddress)
+
+  final def localAddress: IO[Exception, SocketAddress] =
+    IO.syncException(channel.getLocalAddress)
+
+  final def connect[A](attachment: A, socketAddress: SocketAddress): IO[Exception, Unit] =
+    wrap[A, JVoid](h => channel.connect(socketAddress, attachment, h)).void
+
+  final def connect(socketAddress: SocketAddress): IO[Exception, Unit] =
+    wrap[Unit, JVoid](h => channel.connect(socketAddress, (), h)).void
+
+  def read[A](dst: ByteBuffer, timeout: Duration, attachment: A): IO[Exception, Int] =
+    wrap[A, JInteger] { h =>
+      channel.read(dst.buffer, timeout.length, timeout.unit, attachment, h)
+    }.map(_.toInt)
+
+  def read[A](
+    dsts: IList[ByteBuffer],
+    offset: Int,
+    length: Int,
+    timeout: Duration,
+    attachment: A
+  ): IO[Exception, Long] =
+    wrap[A, JLong](
+      h =>
+        channel.read(
+          dsts.map(_.buffer).toList.toArray,
+          offset,
+          length,
+          timeout.length,
+          timeout.unit,
+          attachment,
+          h
+        )
+    ).map(_.toLong)
+}
+
+object AsynchronousSocketChannel {
+
+  def apply(): IO[Exception, AsynchronousSocketChannel] =
+    IO.syncException(JAsynchronousSocketChannel.open())
+      .map(new AsynchronousSocketChannel(_))
+
+  def apply(channelGroup: AsynchronousChannelGroup): IO[Exception, AsynchronousSocketChannel] =
+    IO.syncException(
+        JAsynchronousSocketChannel.open(channelGroup.jChannelGroup)
+      )
+      .map(new AsynchronousSocketChannel(_))
+
+  def apply(asyncSocketChannel: JAsynchronousSocketChannel): AsynchronousSocketChannel =
+    new AsynchronousSocketChannel(asyncSocketChannel)
+}
+
+class AsynchronousChannelGroup(val jChannelGroup: JAsynchronousChannelGroup) {}
+
+object AsynchronousChannelGroup {
+
+  def apply(): IO[Exception, AsynchronousChannelGroup] =
+    ??? // IO.syncException { throw new Exception() }
+}
+
+object AsynchronousChannel {
+  private[nio] def wrap[A, T](op: JCompletionHandler[T, A] => Unit): IO[Exception, T] =
+    IO.async0[Exception, T] { k =>
+      val handler = new JCompletionHandler[T, A] {
+        def completed(result: T, u: A): Unit =
+          k(ExitResult.Completed(result))
+
+        def failed(t: Throwable, u: A): Unit =
+          t match {
+            case e: Exception => k(ExitResult.Failed(e))
+            case _            => k(ExitResult.Terminated(List(t)))
+          }
+      }
+
+      try {
+        op(handler)
+        Async.later[Exception, T]
+      } catch {
+        case e: Exception => Async.now(ExitResult.Failed(e))
+        case t: Throwable => Async.now(ExitResult.Terminated(List(t)))
+      }
+    }
+}
